refactored swagger codegen cli package
