+++ modules/swagger-codegen-cli/src/main/java/io/swagger/codegen/v3/cli/CLIHelper.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen;
+package io.swagger.codegen.v3.cli;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
