+++ modules/swagger-codegen-cli/src/main/java/io/swagger/codegen/v3/cli/cmd/ValidateException.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.cmd;
+package io.swagger.codegen.v3.cli.cmd;
 
 /**
  * Created by takuro on 2017/05/02.
